Move signup email fragment to login library
